requirement.txt
# requirements.txt
mysql-connector-python
bcrypt